No more warning about not defined functions because of wrong header files inclusion. Some variables are now accessible using functions rather than using global variables that can not be accessed through separated files.
Also fix timer returning function type.

+++ src/timer.hpp
@@ -19,7 +19,7 @@
 
 #include <sys/timeb.h>
 
-typedef unsigned long (*TimeFunction)();
+typedef long long int (*TimeFunction)();
 
 /**
  * @brief: A frequency counting clock.
